packages/ciocore/ciocore-8.1.0b1-py2.py3-none-any.whl/tests/test_api_client.py
""" test data

   isort:skip_file
"""
 
import unittest

try:
    from unittest import mock
except ImportError:
    import mock

from ciocore import api_client


class ApiClientTest(unittest.TestCase):
    @staticmethod
    def path_exists_side_effect(arg):
        if "missing" in arg:
            return False
        else:
            return True

    def setUp(self):
        self.env = {"USERPROFILE": "/users/joebloggs", "HOME": "/users/joebloggs"}

        self.api_key_dict = {"api_key": {"client_id": "123", "private_key": "secret123"}}

        patcher = mock.patch("os.path.exists")
        self.mock_exists = patcher.start()
        self.mock_exists.side_effect = ApiClientTest.path_exists_side_effect
        self.addCleanup(patcher.stop)

    def test_create(self):
        ac = api_client.ApiClient()
        self.assertEqual(ac.__class__.__name__, "ApiClient")

    def test_get_standard_creds_path(self):
        with mock.patch.dict("os.environ", self.env):
            fn = api_client.get_creds_path(api_key=False).replace("\\", "/")
            self.assertEqual(fn, "/users/joebloggs/.config/conductor/credentials")

    def test_get_api_key_creds_path(self):
        with mock.patch.dict("os.environ", self.env):
            fn = api_client.get_creds_path(api_key=True).replace("\\", "/")
            self.assertEqual(fn, "/users/joebloggs/.config/conductor/api_key_credentials")

